+++ Sources/LiveKit/Core/Room+MulticastDelegate.swift
@@ -19,16 +19,16 @@ import Foundation
 extension Room: MulticastDelegateProtocol {
     @objc(addDelegate:)
     public func add(delegate: RoomDelegate) {
-        delegates.add(delegate: delegate)
+        _delegates.add(delegate: delegate)
     }
 
     @objc(removeDelegate:)
     public func remove(delegate: RoomDelegate) {
-        delegates.remove(delegate: delegate)
+        _delegates.remove(delegate: delegate)
     }
 
     @objc
     public func removeAllDelegates() {
-        delegates.removeAllDelegates()
+        _delegates.removeAllDelegates()
     }
 }
