+++ Sources/LiveKit/Core/SignalClient.swift
@@ -290,7 +290,7 @@ private extension SignalClient {
             _delegate.notifyAsync { await $0.signalClient(self, didReceiveOffer: sd.toRTCType()) }
 
         case let .trickle(trickle):
-            guard let rtcCandidate = try? Engine.createIceCandidate(fromJsonString: trickle.candidateInit) else {
+            guard let rtcCandidate = try? Room.createIceCandidate(fromJsonString: trickle.candidateInit) else {
                 return
             }
