+++ Sources/LiveKit/Participant/LocalParticipant.swift
@@ -48,7 +48,7 @@ public class LocalParticipant: Participant {
 
         let room = try requireRoom()
 
-        guard let publisher = room.engine.publisher else {
+        guard let publisher = room.publisher else {
             throw LiveKitError(.invalidState, message: "Publisher is nil")
         }
 
@@ -89,7 +89,7 @@ public class LocalParticipant: Participant {
 
                     self.log("[publish] computing encode settings with dimensions: \(dimensions)...")
 
-                    let publishOptions = (publishOptions as? VideoPublishOptions) ?? room._state.options.defaultVideoPublishOptions
+                    let publishOptions = (publishOptions as? VideoPublishOptions) ?? room._state.roomOptions.defaultVideoPublishOptions
                     publishName = publishOptions.name
 
                     let encodings = Utils.computeVideoEncodings(dimensions: dimensions,
@@ -131,7 +131,7 @@ public class LocalParticipant: Participant {
 
                 } else if track is LocalAudioTrack {
                     // additional params for Audio
-                    let publishOptions = (publishOptions as? AudioPublishOptions) ?? room._state.options.defaultAudioPublishOptions
+                    let publishOptions = (publishOptions as? AudioPublishOptions) ?? room._state.roomOptions.defaultAudioPublishOptions
                     publishName = publishOptions.name
 
                     populator.disableDtx = !publishOptions.dtx
@@ -141,7 +141,7 @@ public class LocalParticipant: Participant {
                     self.log("[publish] maxBitrate: \(encoding.maxBitrate)")
 
                     transInit.sendEncodings = [
-                        Engine.createRtpEncodingParameters(encoding: encoding),
+                        Room.createRtpEncodingParameters(encoding: encoding),
                     ]
                 }
 
@@ -154,12 +154,12 @@ public class LocalParticipant: Participant {
             }
 
             // Request a new track to the server
-            let addTrackResult = try await room.engine.signalClient.sendAddTrack(cid: track.mediaTrack.trackId,
-                                                                                 name: publishName ?? track.name,
-                                                                                 type: track.kind.toPBType(),
-                                                                                 source: track.source.toPBType(),
-                                                                                 encryption: room.e2eeManager?.e2eeOptions.encryptionType.toPBType() ?? .none,
-                                                                                 populatorFunc)
+            let addTrackResult = try await room.signalClient.sendAddTrack(cid: track.mediaTrack.trackId,
+                                                                          name: publishName ?? track.name,
+                                                                          type: track.kind.toPBType(),
+                                                                          source: track.source.toPBType(),
+                                                                          encryption: room.e2eeManager?.e2eeOptions.encryptionType.toPBType() ?? .none,
+                                                                          populatorFunc)
 
             log("[Publish] server responded trackInfo: \(addTrackResult.trackInfo)")
 
@@ -184,7 +184,7 @@ public class LocalParticipant: Participant {
                         track._videoCodec = firstVideoCodec
                     }
 
-                    let publishOptions = (publishOptions as? VideoPublishOptions) ?? room._state.options.defaultVideoPublishOptions
+                    let publishOptions = (publishOptions as? VideoPublishOptions) ?? room._state.roomOptions.defaultVideoPublishOptions
                     // if screen share or simulcast is enabled,
                     // degrade resolution by using server's layer switching logic instead of WebRTC's logic
                     if track.source == .screenShareVideo || publishOptions.simulcast {
@@ -201,7 +201,7 @@ public class LocalParticipant: Participant {
                     }
                 }
 
-                try await room.engine.publisherShouldNegotiate()
+                try await room.publisherShouldNegotiate()
                 try Task.checkCancellation()
 
             } catch {
@@ -220,7 +220,7 @@ public class LocalParticipant: Participant {
             delegates.notify(label: { "localParticipant.didPublish \(publication)" }) {
                 $0.participant?(self, didPublishTrack: publication)
             }
-            room.delegates.notify(label: { "localParticipant.didPublish \(publication)" }) {
+            room._delegates.notify(label: { "localParticipant.didPublish \(publication)" }) {
                 $0.room?(room, participant: self, didPublishTrack: publication)
             }
 
@@ -274,7 +274,7 @@ public class LocalParticipant: Participant {
             delegates.notify(label: { "localParticipant.didUnpublish \(publication)" }) {
                 $0.participant?(self, didUnpublishTrack: publication)
             }
-            room.delegates.notify(label: { "room.didUnpublish \(publication)" }) {
+            room._delegates.notify(label: { "room.didUnpublish \(publication)" }) {
                 $0.room?(room, participant: self, didUnpublishTrack: publication)
             }
         }
@@ -288,11 +288,11 @@ public class LocalParticipant: Participant {
         }
 
         // Wait for track to stop (if required)
-        if room._state.options.stopLocalTrackOnUnpublish {
+        if room._state.roomOptions.stopLocalTrackOnUnpublish {
             try await track.stop()
         }
 
-        if let publisher = room.engine.publisher, let sender = track.rtpSender {
+        if let publisher = room.publisher, let sender = track.rtpSender {
             // Remove all simulcast senders...
             for simulcastSender in track._simulcastRtpSenders.values {
                 try await publisher.remove(track: simulcastSender)
@@ -300,7 +300,7 @@ public class LocalParticipant: Participant {
             // Remove main sender...
             try await publisher.remove(track: sender)
             // Mark re-negotiation required...
-            try await room.engine.publisherShouldNegotiate()
+            try await room.publisherShouldNegotiate()
         }
 
         try await track.onUnpublish()
@@ -317,7 +317,7 @@ public class LocalParticipant: Participant {
     @objc
     public func publish(data: Data, options: DataPublishOptions? = nil) async throws {
         let room = try requireRoom()
-        let options = options ?? room._state.options.defaultDataPublishOptions
+        let options = options ?? room._state.roomOptions.defaultDataPublishOptions
 
         let userPacket = Livekit_UserPacket.with {
             $0.participantSid = self.sid
@@ -326,7 +326,7 @@ public class LocalParticipant: Participant {
             $0.topic = options.topic ?? ""
         }
 
-        try await room.engine.send(userPacket: userPacket, kind: options.reliable ? .reliable : .lossy)
+        try await room.send(userPacket: userPacket, kind: options.reliable ? .reliable : .lossy)
     }
 
     /**
@@ -368,7 +368,7 @@ public class LocalParticipant: Participant {
 
         // TODO: Revert internal state on failure
         let room = try requireRoom()
-        try await room.engine.signalClient.sendUpdateLocalMetadata(metadata, name: name)
+        try await room.signalClient.sendUpdateLocalMetadata(metadata, name: name)
     }
 
     /// Sets and updates the name of the local participant.
@@ -383,15 +383,15 @@ public class LocalParticipant: Participant {
 
         // TODO: Revert internal state on failure
         let room = try requireRoom()
-        try await room.engine.signalClient.sendUpdateLocalMetadata(metadata ?? "", name: name)
+        try await room.signalClient.sendUpdateLocalMetadata(metadata ?? "", name: name)
     }
 
     func sendTrackSubscriptionPermissions() async throws {
         let room = try requireRoom()
-        guard room.engine._state.connectionState == .connected else { return }
+        guard room._state.connectionState == .connected else { return }
 
-        try await room.engine.signalClient.sendUpdateSubscriptionPermission(allParticipants: allParticipantsAllowed,
-                                                                            trackPermissions: trackPermissions)
+        try await room.signalClient.sendUpdateSubscriptionPermission(allParticipants: allParticipantsAllowed,
+                                                                     trackPermissions: trackPermissions)
     }
 
     func _set(subscribedQualities qualities: [Livekit_SubscribedQuality], forTrackSid trackSid: String) {
@@ -411,7 +411,7 @@ public class LocalParticipant: Participant {
             delegates.notify(label: { "participant.didUpdatePermissions: \(newValue)" }) {
                 $0.participant?(self, didUpdatePermissions: newValue)
             }
-            room.delegates.notify(label: { "room.didUpdatePermissions: \(newValue)" }) {
+            room._delegates.notify(label: { "room.didUpdatePermissions: \(newValue)" }) {
                 $0.room?(room, participant: self, didUpdatePermissions: newValue)
             }
         }
@@ -511,31 +511,31 @@ public extension LocalParticipant {
         } else if enabled {
             // Try to create a new track
             if source == .camera {
-                let localTrack = LocalVideoTrack.createCameraTrack(options: (captureOptions as? CameraCaptureOptions) ?? room._state.options.defaultCameraCaptureOptions,
-                                                                   reportStatistics: room._state.options.reportRemoteTrackStatistics)
+                let localTrack = LocalVideoTrack.createCameraTrack(options: (captureOptions as? CameraCaptureOptions) ?? room._state.roomOptions.defaultCameraCaptureOptions,
+                                                                   reportStatistics: room._state.roomOptions.reportRemoteTrackStatistics)
                 return try await publish(videoTrack: localTrack, publishOptions: publishOptions as? VideoPublishOptions)
             } else if source == .microphone {
-                let localTrack = LocalAudioTrack.createTrack(options: (captureOptions as? AudioCaptureOptions) ?? room._state.options.defaultAudioCaptureOptions,
-                                                             reportStatistics: room._state.options.reportRemoteTrackStatistics)
+                let localTrack = LocalAudioTrack.createTrack(options: (captureOptions as? AudioCaptureOptions) ?? room._state.roomOptions.defaultAudioCaptureOptions,
+                                                             reportStatistics: room._state.roomOptions.reportRemoteTrackStatistics)
                 return try await publish(audioTrack: localTrack, publishOptions: publishOptions as? AudioPublishOptions)
             } else if source == .screenShareVideo {
                 #if os(iOS)
                     let localTrack: LocalVideoTrack
-                    let options = (captureOptions as? ScreenShareCaptureOptions) ?? room._state.options.defaultScreenShareCaptureOptions
-                    if options.useBroadcastExtension {
+                    let roomOptions = (captureOptions as? ScreenShareCaptureOptions) ?? room._state.roomOptions.defaultScreenShareCaptureOptions
+                    if roomOptions.useBroadcastExtension {
                         let screenShareExtensionId = Bundle.main.infoDictionary?[BroadcastScreenCapturer.kRTCScreenSharingExtension] as? String
                         await RPSystemBroadcastPickerView.show(for: screenShareExtensionId, showsMicrophoneButton: false)
-                        localTrack = LocalVideoTrack.createBroadcastScreenCapturerTrack(options: options)
+                        localTrack = LocalVideoTrack.createBroadcastScreenCapturerTrack(roomOptions: roomOptions)
                     } else {
-                        localTrack = LocalVideoTrack.createInAppScreenShareTrack(options: options)
+                        localTrack = LocalVideoTrack.createInAppScreenShareTrack(roomOptions: roomOptions)
                     }
                     return try await publish(videoTrack: localTrack, publishOptions: publishOptions as? VideoPublishOptions)
                 #elseif os(macOS)
                     if #available(macOS 12.3, *) {
                         let mainDisplay = try await MacOSScreenCapturer.mainDisplaySource()
                         let track = LocalVideoTrack.createMacOSScreenShareTrack(source: mainDisplay,
-                                                                                options: (captureOptions as? ScreenShareCaptureOptions) ?? room._state.options.defaultScreenShareCaptureOptions,
-                                                                                reportStatistics: room._state.options.reportRemoteTrackStatistics)
+                                                                                roomOptions: (captureOptions as? ScreenShareCaptureOptions) ?? room._state.roomOptions.defaultScreenShareCaptureOptions,
+                                                                                reportStatistics: room._state.roomOptions.reportRemoteTrackStatistics)
                         return try await publish(videoTrack: track, publishOptions: publishOptions as? VideoPublishOptions)
                     }
                 #endif
@@ -567,9 +567,9 @@ extension LocalParticipant {
             throw LiveKitError(.invalidState, message: "Attempted to publish a non-backup video codec as backup")
         }
 
-        let publisher = try room.engine.requirePublisher()
+        let publisher = try room.requirePublisher()
 
-        let publishOptions = (track.publishOptions as? VideoPublishOptions) ?? room._state.options.defaultVideoPublishOptions
+        let publishOptions = (track.publishOptions as? VideoPublishOptions) ?? room._state.roomOptions.defaultVideoPublishOptions
 
         // Should be already resolved...
         let dimensions = try await track.capturer.dimensionsCompleter.wait()
@@ -595,10 +595,10 @@ extension LocalParticipant {
         let sender = transceiver.sender
 
         // Request a new track to the server
-        let addTrackResult = try await room.engine.signalClient.sendAddTrack(cid: sender.senderId,
-                                                                             name: track.name,
-                                                                             type: track.kind.toPBType(),
-                                                                             source: track.source.toPBType())
+        let addTrackResult = try await room.signalClient.sendAddTrack(cid: sender.senderId,
+                                                                      name: track.name,
+                                                                      type: track.kind.toPBType(),
+                                                                      source: track.source.toPBType())
         {
             $0.sid = localTrackPublication.sid
             $0.simulcastCodecs = [
@@ -618,7 +618,7 @@ extension LocalParticipant {
         // Attach multi-codec sender...
         track._simulcastRtpSenders[videoCodec] = sender
 
-        try await room.engine.publisherShouldNegotiate()
+        try await room.publisherShouldNegotiate()
     }
 }
