+++ Sources/LiveKit/Protocols/EngineDelegate.swift
@@ -17,11 +17,3 @@
 import Foundation
 
 @_implementationOnly import WebRTC
-
-protocol EngineDelegate: AnyObject {
-    func engine(_ engine: Engine, didMutateState state: Engine.State, oldState: Engine.State) async
-    func engine(_ engine: Engine, didUpdateSpeakers speakers: [Livekit_SpeakerInfo]) async
-    func engine(_ engine: Engine, didAddTrack track: LKRTCMediaStreamTrack, rtpReceiver: LKRTCRtpReceiver, stream: LKRTCMediaStream) async
-    func engine(_ engine: Engine, didRemoveTrack track: LKRTCMediaStreamTrack) async
-    func engine(_ engine: Engine, didReceiveUserPacket packet: Livekit_UserPacket) async
-}
