+++ Sources/LiveKit/Support/Utils.swift
@@ -216,11 +216,11 @@ class Utils {
         if let videoCodec, videoCodec.isSVC {
             // SVC mode
             logger.log("Using SVC mode", type: Utils.self)
-            return [Engine.createRtpEncodingParameters(encoding: encoding, scalabilityMode: .L3T3_KEY)]
+            return [Room.createRtpEncodingParameters(encoding: encoding, scalabilityMode: .L3T3_KEY)]
         } else if !publishOptions.simulcast {
             // Not-simulcast mode
             logger.log("Simulcast not enabled", type: Utils.self)
-            return [Engine.createRtpEncodingParameters(encoding: encoding)]
+            return [Room.createRtpEncodingParameters(encoding: encoding)]
         }
 
         // Continue to simulcast encoding computation...
