+++ Sources/LiveKit/Track/Capturers/BufferCapturer.swift
@@ -29,7 +29,7 @@ import Foundation
 /// since dimensions must be resolved at the time of publishing (to compute video parameters).
 ///
 public class BufferCapturer: VideoCapturer {
-    private let capturer = Engine.createVideoCapturer()
+    private let capturer = Room.createVideoCapturer()
 
     /// The ``BufferCaptureOptions`` used for this capturer.
     public var options: BufferCaptureOptions
@@ -88,7 +88,7 @@ public extension LocalVideoTrack {
                                   options: BufferCaptureOptions = BufferCaptureOptions(),
                                   reportStatistics: Bool = false) -> LocalVideoTrack
     {
-        let videoSource = Engine.createVideoSource(forScreenShare: source == .screenShareVideo)
+        let videoSource = Room.createVideoSource(forScreenShare: source == .screenShareVideo)
         let capturer = BufferCapturer(delegate: videoSource, options: options)
         return LocalVideoTrack(name: name,
                                source: source,
