+++ Sources/LiveKit/Track/Capturers/InAppCapturer.swift
@@ -24,7 +24,7 @@ import Foundation
 
 @available(macOS 11.0, iOS 11.0, *)
 public class InAppScreenCapturer: VideoCapturer {
-    private let capturer = Engine.createVideoCapturer()
+    private let capturer = Room.createVideoCapturer()
     private var options: ScreenShareCaptureOptions
 
     init(delegate: LKRTCVideoCapturerDelegate, options: ScreenShareCaptureOptions) {
@@ -82,7 +82,7 @@ public extension LocalVideoTrack {
                                             options: ScreenShareCaptureOptions = ScreenShareCaptureOptions(),
                                             reportStatistics: Bool = false) -> LocalVideoTrack
     {
-        let videoSource = Engine.createVideoSource(forScreenShare: true)
+        let videoSource = Room.createVideoSource(forScreenShare: true)
         let capturer = InAppScreenCapturer(delegate: videoSource, options: options)
         return LocalVideoTrack(name: name,
                                source: .screenShareVideo,
