+++ Sources/LiveKit/TrackPublications/LocalTrackPublication.swift
@@ -106,7 +106,7 @@ extension LocalTrackPublication {
         let parameters = sender.parameters
 
         guard let participant, let room = participant._room else { return }
-        let publishOptions = (track.publishOptions as? VideoPublishOptions) ?? room._state.options.defaultVideoPublishOptions
+        let publishOptions = (track.publishOptions as? VideoPublishOptions) ?? room._state.roomOptions.defaultVideoPublishOptions
 
         // re-compute encodings
         let encodings = Utils.computeVideoEncodings(dimensions: dimensions,
@@ -145,7 +145,7 @@ extension LocalTrackPublication {
         Task.detached {
             let participant = try await self.requireParticipant()
             let room = try participant.requireRoom()
-            try await room.engine.signalClient.sendUpdateVideoLayers(trackSid: track.sid!, layers: layers)
+            try await room.signalClient.sendUpdateVideoLayers(trackSid: track.sid!, layers: layers)
         }
     }
 }
